offload xlsx import parsing to a comlink worker
Profiled imports blocked the main thread ~9s in one task (mostly the
luckyexcel parse). The parse/decorate/compact pipeline now lives in a
pure module shared by a blob-inlined dsheet worker and the main-thread
fallback, so the two paths cannot diverge. The worker is a generic
comlink task API (xlsx parse is its first method), spawned lazily and
terminated after 30s idle. Toasts, sheet-id generation and the ydoc
write stay on the main thread.
