+++ package.json
@@ -2,7 +2,7 @@
   "name": "@fileverse-dev/dsheet",
   "private": false,
   "description": "DSheet",
-  "version": "4.0.3-import-compaction.0",
+  "version": "4.0.3-import-compaction.1",
   "main": "dist/index.es.js",
   "module": "dist/index.es.js",
   "exports": {
@@ -69,6 +69,7 @@
     "@tippyjs/react": "^4.2.6",
     "@ucans/ucans": "^0.12.0",
     "classnames": "^2.5.1",
+    "comlink": "^4.4.2",
     "dayjs": "^1.11.0",
     "exceljs": "^4.4.0",
     "immer": "^9.0.12",
